+++ packages/monaco-editor-wrapper/src/index.ts
@@ -14,7 +14,7 @@ import * as vscode from 'vscode';
 
 import { getMonacoCss } from './generated/css.js';
 
-import { MonacoLanguageClient, CloseAction, ErrorAction, MonacoServices, MessageTransports, MessageWriter, MessageReader } from 'monaco-languageclient';
+import { MonacoLanguageClient, CloseAction, ErrorAction, MonacoServices, MessageTransports } from 'monaco-languageclient';
 import { toSocket, WebSocketMessageReader, WebSocketMessageWriter } from 'vscode-ws-jsonrpc';
 import { BrowserMessageReader, BrowserMessageWriter } from 'vscode-languageserver-protocol/browser.js';
 import normalizeUrl from 'normalize-url';
@@ -211,11 +211,6 @@ export class CodeEditorConfig {
     }
 }
 
-export type CommunitcationConfig = {
-    reader: MessageReader,
-    writer: MessageWriter
-};
-
 export class MonacoEditorLanguageClientWrapper {
 
     private editor: monaco.editor.IStandaloneCodeEditor | undefined;
@@ -223,7 +218,7 @@ export class MonacoEditorLanguageClientWrapper {
     private editorConfig: CodeEditorConfig = new CodeEditorConfig();
     private languageClient: MonacoLanguageClient | undefined;
     private worker: Worker | undefined;
-    private communitcationConfig: CommunitcationConfig | undefined;
+    private messageTransports: MessageTransports | undefined;
 
     private dispatchEvent: ((event: Event) => boolean) | undefined;
 
@@ -265,15 +260,15 @@ export class MonacoEditorLanguageClientWrapper {
         monaco.editor.setTheme(this.editorConfig.getTheme());
     }
 
-    setWorker(worker: Worker, communitcationConfig?: CommunitcationConfig) {
+    setWorker(worker: Worker, messageTransports?: MessageTransports) {
         this.worker = worker;
-        if (communitcationConfig) {
-            this.communitcationConfig = communitcationConfig;
+        if (messageTransports) {
+            this.messageTransports = messageTransports;
         }
     }
 
-    getCommunitcationConfig(): CommunitcationConfig | undefined {
-        return this.communitcationConfig;
+    getMessageTransports(): MessageTransports | undefined {
+        return this.messageTransports;
     }
 
     isStarted(): boolean {
@@ -485,11 +480,11 @@ export class MonacoEditorLanguageClientWrapper {
 
                 webSocket.onopen = () => {
                     const socket = toSocket(webSocket);
-                    this.communitcationConfig = {
+                    this.messageTransports = {
                         reader: new WebSocketMessageReader(socket),
                         writer: new WebSocketMessageWriter(socket)
                     };
-                    this.handleLanguageClientStart(this.communitcationConfig, resolve, reject);
+                    this.handleLanguageClientStart(this.messageTransports, resolve, reject);
                 };
             } else {
                 const workerConfigOptions = lcConfigOptions as WorkerConfigOptions;
@@ -499,26 +494,23 @@ export class MonacoEditorLanguageClientWrapper {
                         name: workerConfigOptions.workerName,
                     });
                 }
-                if (!this.communitcationConfig) {
-                    this.communitcationConfig = {
+                if (!this.messageTransports) {
+                    this.messageTransports = {
                         reader: new BrowserMessageReader(this.worker),
                         writer: new BrowserMessageWriter(this.worker)
                     };
                 }
-                this.handleLanguageClientStart(this.communitcationConfig, resolve, reject);
+                this.handleLanguageClientStart(this.messageTransports, resolve, reject);
             }
         });
     }
 
-    private async handleLanguageClientStart(communitcationConfig: CommunitcationConfig,
+    private async handleLanguageClientStart(messageTransports: MessageTransports,
         resolve: (value: string) => void,
         reject: (reason?: unknown) => void) {
 
-        this.languageClient = this.createLanguageClient({
-            reader: communitcationConfig.reader,
-            writer: communitcationConfig.writer
-        });
-        communitcationConfig.reader.onClose(() => this.languageClient?.stop());
+        this.languageClient = this.createLanguageClient(messageTransports);
+        messageTransports.reader.onClose(() => this.languageClient?.stop());
 
         await this.languageClient.start()
             .then(() => {